// src/Common/enums.ts
export enum FEnums {
    AssetManagement = "103",
    Settings = "995",
    MySettings = "95",
    Import = "945",
    Export = "950",
    InventoryManagement = "353",
    InventoryConfiguration = "356",
    AssetConfiguration = "106",
    DeviceManagement = "253",
    DeviceConfiguration = "256",
    NewSession = "12",
    AssetPowerCabling = "153",
    AssetNetworkCabling = "156",
    InventoryImport = "365",
    InventoryExport = "368",
    DevicePowerCabling = "259",
    DeviceNetworkCabling = "262",
    DeviceImport = "265",
    DeviceExport = "268",
    ImportCables = "373",
    ExportCables = "376",
    Discover = "306",
    Monitor = "309",
    IntelDCMConsole = "303",
    Configure = "312",
    // PlatformOS = "353",
    // Applications = "356",
    PhysicalCompute = "403",
    VirtualCompute = "406",
    CloudCompute = "409",
    ComputeAlerts = "412",
    ModelBusinessService = "453",
    BusinessServiceAlerts = "459",

    TrendsAndForecast = "509",
    FilterCapacity = "512",
    OptimizeCapacity = "515",
    ReportBuilder = "653",
    Optimize = "903",
    Stop = "921",
    Start = "924",
    ManageAuditSessions = "603",
    AuditDataCenter = "606",
    InventoryReconciliation = "609",
    BackgroundTasks = "18",
    Reminders = "21",
    UserProfile = "8",
    Documentation = "6",
    DataCenterHierarchy = "905",
    InventoryStoreHierarchy = "910",
    ManageBusinessService = "456",
    Home = "990",
    MigrateDB = "915",
    PowerBI = "506",
    AnalyticsAndCharts = "503",
    close = "4",
    FloorLayout = "258",
    StopNetZoom = "909",
    StartNetZoom = "906",
}